refactor(utils): rename debug utils files
