+++ planning/behavior_path_planner/CMakeLists.txt
@@ -50,9 +50,9 @@ ament_auto_add_library(behavior_path_planner_node SHARED
   src/utils/drivable_area_expansion/expansion.cpp
   src/utils/geometric_parallel_parking/geometric_parallel_parking.cpp
   src/utils/occupancy_grid_based_collision_detector/occupancy_grid_based_collision_detector.cpp
-  src/marker_util/debug_utilities.cpp
-  src/marker_util/avoidance/debug.cpp
-  src/marker_util/lane_change/debug.cpp
+  src/marker_utils/utils.cpp
+  src/marker_utils/avoidance/debug.cpp
+  src/marker_utils/lane_change/debug.cpp
 )
 
 target_include_directories(behavior_path_planner_node SYSTEM PUBLIC
