+++ planning/behavior_path_planner/src/utils/safety_check.cpp
@@ -14,7 +14,7 @@
 
 #include "behavior_path_planner/utils/safety_check.hpp"
 
-#include "behavior_path_planner/marker_util/debug_utilities.hpp"
+#include "behavior_path_planner/marker_utils/utils.hpp"
 #include "interpolation/linear_interpolation.hpp"
 #include "motion_utils/trajectory/trajectory.hpp"
 #include "object_recognition_utils/predicted_path_utils.hpp"
