+++ planning/behavior_path_planner/test/test_safety_check.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "behavior_path_planner/marker_util/debug_utilities.hpp"
+#include "behavior_path_planner/marker_utils/utils.hpp"
 #include "behavior_path_planner/utils/safety_check.hpp"
 #include "tier4_autoware_utils/tier4_autoware_utils.hpp"
